refactor: share task field validation across backends
Move the backend-agnostic normalize/compare helpers out of the markdown
store into src/backends/validate.ts so the github backend reuses them
instead of growing a second copy. No behavior change.
