+++ src/backends/markdown.ts
@@ -10,7 +10,6 @@ import { AxiError } from "../errors.js";
 import { validateDependencyId, validateId } from "../id.js";
 import type {
   Dep,
-  Hold,
   State,
   Task,
   TaskInput,
@@ -21,7 +20,6 @@ import type {
   TaskUpdateResult,
   TransitionOpts,
 } from "../model.js";
-import { HOLD_KINDS } from "../model.js";
 import {
   PUBLIC_FOLLOWUP_KIND,
   assertPublicFollowupMutation,
@@ -41,12 +39,24 @@ import type {
 } from "../store.js";
 import { atomicWrite, readFileSafe, withLock, withLocks } from "./lock.js";
 import {
+  addBodyLine,
+  bodyHasLine,
+  normalizeDate,
+  normalizeDep,
+  normalizeHold,
+  normalizePriority,
+  normalizeTagValue,
+  normalizeTitle,
+  normalizeTypedLink,
+  sameHold,
+  sameMeta,
+} from "./validate.js";
+import {
   type BacklogDoc,
   type Entry,
   type Section,
   type TaskEntry,
   deriveLinks,
-  extractTags,
   parseBacklog,
   renderBacklog,
   renderTaskLines,
@@ -68,10 +78,6 @@ const HEADERS: Record<State, string> = {
   queued: "## Queued",
   done: "## Done",
 };
-const DATE_RE = /^\d{4}-\d{2}-\d{2}$/;
-const DEP_REASON_EDGE_MARKER_RE =
-  /(?:^|\s)(?:blocked-by|parent|discovered-from):\s/;
-
 interface LoadedBacklogDoc {
   doc: BacklogDoc;
   source: string | undefined;
@@ -96,184 +102,10 @@ function errno(error: unknown): string {
     : "UNKNOWN";
 }
 
-function normalizeTitle(title: string): string {
-  if (/[\r\n]/.test(title)) {
-    throw new AxiError("Task title must be a single line", "VALIDATION_ERROR");
-  }
-  const trimmed = title.trim();
-  if (trimmed === "") {
-    throw new AxiError("Task title must not be empty", "VALIDATION_ERROR");
-  }
-  if (extractTags(trimmed).title !== trimmed) {
-    throw new AxiError(
-      "Task title must not end with canonical task tags",
-      "VALIDATION_ERROR",
-    );
-  }
-  return trimmed;
-}
-
-function normalizeTagValue(
-  value: string | undefined,
-  field: "kind" | "repo",
-): string | undefined {
-  if (value === undefined) return undefined;
-  if (/[()\r\n]/.test(value)) {
-    throw new AxiError(
-      `Task ${field} must be a single line without parentheses`,
-      "VALIDATION_ERROR",
-    );
-  }
-  const trimmed = value.trim();
-  return trimmed === "" ? undefined : trimmed;
-}
-
-function normalizeLinkUrl(url: string): string {
-  if (/[\r\n]/.test(url)) {
-    throw new AxiError("Task link must be a single line", "VALIDATION_ERROR");
-  }
-  const trimmed = url.trim();
-  if (trimmed === "") {
-    throw new AxiError("Task link must not be empty", "VALIDATION_ERROR");
-  }
-  return trimmed;
-}
-
-function normalizeTypedLink(link: TaskLink): TaskLink {
-  const url = normalizeLinkUrl(link.url);
-  const derived = deriveLinks(url);
-  if (
-    !derived.some(
-      (candidate) => candidate.kind === link.kind && candidate.url === url,
-    )
-  ) {
-    const expected =
-      link.kind === "pr"
-        ? "an http(s) pull request URL ending in /pull/<number>"
-        : link.kind === "report"
-          ? "a data/<id>/report.md path"
-          : "an http(s) URL";
-    throw new AxiError(
-      `Task ${link.kind} link must be ${expected}`,
-      "VALIDATION_ERROR",
-    );
-  }
-  return { kind: link.kind, url };
-}
-
-function normalizePriority(priority: number | undefined): number | undefined {
-  if (priority === undefined) return undefined;
-  if (!Number.isInteger(priority) || priority < 0 || priority > 4) {
-    throw new AxiError(
-      "Task priority must be an integer 0-4",
-      "VALIDATION_ERROR",
-    );
-  }
-  return priority;
-}
-
-function normalizeHold(hold: Hold | undefined): Hold | undefined {
-  if (hold === undefined) return undefined;
-  if (/[\r\n()]/.test(hold.reason)) {
-    throw new AxiError(
-      "Task hold reason must be a single line without parentheses",
-      "VALIDATION_ERROR",
-    );
-  }
-  const reason = hold.reason.trim();
-  if (reason === "") {
-    throw new AxiError(
-      "Task hold reason must not be empty",
-      "VALIDATION_ERROR",
-    );
-  }
-  const normalized: Hold = { reason };
-  if (hold.kind !== undefined) {
-    if (!(HOLD_KINDS as readonly string[]).includes(hold.kind)) {
-      throw new AxiError(
-        `Task hold kind must be one of ${HOLD_KINDS.join(", ")}`,
-        "VALIDATION_ERROR",
-      );
-    }
-    normalized.kind = hold.kind;
-  }
-  if (hold.until !== undefined) {
-    normalized.until = normalizeDate(hold.until, "hold-until date");
-  }
-  return normalized;
-}
-
-function normalizeDate(value: string, field: string): string {
-  if (!DATE_RE.test(value)) {
-    throw new AxiError(`Task ${field} must be YYYY-MM-DD`, "VALIDATION_ERROR");
-  }
-  return value;
-}
-
-function normalizeDepReason(reason: string | undefined): string | undefined {
-  if (reason === undefined) return undefined;
-  if (/[\r\n]/.test(reason)) {
-    throw new AxiError(
-      "Task dependency reason must be a single line",
-      "VALIDATION_ERROR",
-    );
-  }
-  const trimmed = reason.trim();
-  if (DEP_REASON_EDGE_MARKER_RE.test(trimmed)) {
-    throw new AxiError(
-      "Task dependency reason must not contain dependency markers",
-      "VALIDATION_ERROR",
-    );
-  }
-  return trimmed === "" ? undefined : trimmed;
-}
-
-function normalizeDep(ownerId: string, dep: Dep): Dep {
-  const reason = normalizeDepReason(dep.reason);
-  const checked: Dep = { ...dep, id: validateDependencyId(dep.id) };
-  if (reason === undefined) {
-    delete checked.reason;
-  } else {
-    checked.reason = reason;
-  }
-  if (checked.id === ownerId) {
-    throw new AxiError("A task cannot block itself", "VALIDATION_ERROR");
-  }
-  return checked;
-}
-
 function appendTitleLink(title: string, link: TaskLink): string {
   const { url } = normalizeTypedLink(link);
   if (deriveLinks(title).some((link) => link.url === url)) return title;
   return normalizeTitle(`${title} ${url}`);
-}
-
-function bodyHasLine(body: string | undefined, line: string): boolean {
-  return body?.split("\n").includes(line) ?? false;
-}
-
-function addBodyLine(body: string | undefined, line: string): string {
-  return body ? `${body}\n${line}` : line;
-}
-
-function sameHold(left: Hold | undefined, right: Hold | undefined): boolean {
-  return (
-    left?.reason === right?.reason &&
-    left?.kind === right?.kind &&
-    left?.until === right?.until
-  );
-}
-
-function sameMeta(
-  left: Record<string, unknown> | undefined,
-  right: Record<string, unknown> | undefined,
-): boolean {
-  const leftKeys = Object.keys(left ?? {});
-  const rightKeys = Object.keys(right ?? {});
-  return (
-    leftKeys.length === rightKeys.length &&
-    leftKeys.every((key) => Object.is(left?.[key], right?.[key]))
-  );
 }
 
 function taskToInput(task: Task): TaskInput {
